Initial proposition of two steps active messages. Each active message now has (1) a function to run (2) an optional function that returns a pointer to a buffer. This allows to send a buffer along the usual arguments. That buffer will be Isent as is, without any extra copies. On the receiver side, we guarantee that the buffer has arrived (and is store in the user prodived location) before actually running the active function.

+++ tasktorrent/src/active_messages.cpp
@@ -4,8 +4,8 @@
 
 namespace ttor {
 
-int ActiveMsgBase::get_id() const { return id_; }
-ActiveMsgBase::ActiveMsgBase(int id) : id_(id) {}
+size_t ActiveMsgBase::get_id() const { return id_; }
+ActiveMsgBase::ActiveMsgBase(size_t id) : id_(id) {}
 ActiveMsgBase::~ActiveMsgBase(){}
 
 }
